# Collect basic request metrics and expose using GoLang's expvar