--- a/app.go
+++ b/app.go
@@ -3,6 +3,7 @@ package main
 import (
 	"bytes"
 	"encoding/json"
+	"expvar"
 	"flag"
 	"fmt"
 	"io"
@@ -31,11 +32,17 @@ const (
 var cert string
 var key string
 var port string
+var metricsEnabled bool
+
+var totalRequestCount expvarInt
+var concurrentRequestCount expvarInt
+var maxConcurrentRequestCount expvarMaxInt
 
 func init() {
 	flag.StringVar(&cert, "cert", "", "give me a certificate")
 	flag.StringVar(&key, "key", "", "give me a key")
 	flag.StringVar(&port, "port", "80", "give me a port number")
+	flag.BoolVar(&metricsEnabled, "metrics", false, "enable collecting metrics")
 }
 
 var upgrader = websocket.Upgrader{
@@ -45,13 +52,14 @@ var upgrader = websocket.Upgrader{
 
 func main() {
 	flag.Parse()
+	publishExpvarMetrics()
 
-	http.HandleFunc("/data", dataHandler)
-	http.HandleFunc("/echo", echoHandler)
-	http.HandleFunc("/bench", benchHandler)
-	http.HandleFunc("/", whoamiHandler)
-	http.HandleFunc("/api", apiHandler)
-	http.HandleFunc("/health", healthHandler)
+	http.HandleFunc("/data", wrapHandlerIfNeeded(dataHandler))
+	http.HandleFunc("/echo", wrapHandlerIfNeeded(echoHandler))
+	http.HandleFunc("/bench", wrapHandlerIfNeeded(benchHandler))
+	http.HandleFunc("/", wrapHandlerIfNeeded(whoamiHandler))
+	http.HandleFunc("/api", wrapHandlerIfNeeded(apiHandler))
+	http.HandleFunc("/health", wrapHandlerIfNeeded(healthHandler))
 
 	fmt.Println("Starting up on port " + port)
 
@@ -261,4 +269,35 @@ func fillContent(length int64) io.ReadSeeker {
 	}
 
 	return bytes.NewReader(b)
+}
+
+func publishExpvarMetrics() {
+	if metricsEnabled {
+		metrics := expvar.Map{}
+		metrics.Set("totalRequestCount", &totalRequestCount)
+		metrics.Set("concurrentRequestCount", &concurrentRequestCount)
+		metrics.Set("maxConcurrentRequestCount", &maxConcurrentRequestCount)
+		expvar.Publish("metrics", &metrics)
+	}
+}
+
+func wrapHandlerIfNeeded(handler http.HandlerFunc) http.HandlerFunc {
+	if !metricsEnabled {
+		return handler
+	}
+	return func(writer http.ResponseWriter, request *http.Request) {
+		onHandleRequestStart()
+		handler(writer, request)
+		onHandleRequestFinish()
+	}
+}
+
+func onHandleRequestStart() {
+	count := concurrentRequestCount.Add(1)
+	totalRequestCount.Add(1)
+	maxConcurrentRequestCount.Update(count)
+}
+
+func onHandleRequestFinish() {
+	concurrentRequestCount.Add(-1)
 }