@yinonavraham 2020-11-25
Can someone please review this PR, approve or reject? Any feedback is welcome... I really don't like to keep PRs open for long time - they don't age well...